// src/enum/options.enum.ts
export enum FetchOptions {
  COUNTRY = "countries",
  COUNTRY_GCP = "Country",
  INDUSTRY = "Industry",
  FILTER_CONFIG = "filterConfig",
  SALESACTIVITIES = "SalesActivities",
  CAMPAIGN = "CampaignName",
  CAMPAIGN_CODE = "campaignCode",
  DELIVERYMODAL = "DeliveryModel",
  MARKETINGACTIVITY = "MarketingActivityChannel",
  MARKETINGUSECASE = "MarketingActivityUseCase",
  PRIMARYNEED = "PartnerPrimaryNeed",
  COMPETITIVETRACKING = "CompetitiveTracking",
  USECASE = "UseCase",
  STAGE = "Stage",
  FILTER_STAGE = "Stage",
  CLOSEDLOSTREASON = "ClosedLostReason",
  CURRENCYVALUE = "CustomerSoftwareValueCurrency",
  CURRENCY = "Currency",
  PURCHASE_INTENT = "CustomerMarketplacePurchaseIntent",
  SOLUTION_AREA = "SolutionArea",
  SOLUTION_PLAY = "SolutionPlay",
  HELP_TYPE = "HelpType",
  PARTNER_ROLE = "PartnerRole",
  PROCUREMENTTYPE = "ProcurementType",
  REJECTIONREASON = "RejectionReason",
  STATE = "State",
  STATES = "States",
  STATUS = "Status",
  NEED = "need",
  AUTHORITY = "authority",
  BUDGET = "budget",
  TIME_LINE = "timeline",
  PROVIDER = "provider",
  DECISION_PHASE = "DecisionPhase",
  CONTRACT_VEHICLE = "ContractVehicle",
  REGION = "Region",
  SUPPORT_LEVEL = "SupportLevel",
  PRODUCT_FAMILY = "ProductFamily",
  PARTNER_ENGAGEMENT = "PartnerEngagementType",
  PARTNER_CONNECT = "partnerconnections",
  CLOSING_REASON = "ClosingReason",
  SOURCE = "Source",
}

export enum DropdownOptions {
  COUNTRY = "countries",
  COUNTRY_GCP = "countriesGcp",
  COUNTRY_AMP = "countriesAmp",
  INDUSTRY = "industry",
  INDUSTRY_GCP = "industryGcp",
  PRODUCTS = "AWSProducts",
  SALESACTIVITIES = "salesActivites",
  CAMPAIGN = "apnPrograms",
  FILTER_CONFIG = "filterConfig",
  AMP_SOLUTIONS = "ampSolutions",
  DELIVERYMODAL = "deliveryModel",
  DELIVERY_MODAL_GCP = "deliveryModelGcp",
  MARKETINGACTIVITY = "marketingActivityChannel",
  SOULTIONS = "solutions",
  MARKETINGUSECASE = "marketingUseCase",
  PRIMARYNEED = "specificAWSCoSellNeeds",
  COMPETITIVETRACKING = "competitiveTracking",
  USECASE = "useCase",
  STAGE = "stage",
  STAGE_GCP = "stageGcp",
  COUNTRY_STATE_GCP = "countryStateGcp",
  HELP_TYPE = "helpType",
  SOLUTION_AREA = "solutionArea",
  PARTNER_ROLE = "partnerRole",
  SOLUTION_PLAY = "solutionPlay",
  CLOSEDLOSTREASON = "closedLostReason",
  CURRENCYVALUE = "currencyValue",
  CURRENCY = "currency",
  CURRENCY_AMP = "currencyAmp",
  PROCUREMENTTYPE = "procurementType",
  REJECTIONREASON = "RejectionReason",
  STATE = "state",
  STATE_GCP = "stateGcp",
  STATE_AMP = "stateAmp",
  STATUS = "status",
  PROVIDER = "provider",
  PRIVATE_PROVIDER = "privateProvider",
  REFERRAL_TYPE = "referralType",
  COSELL_PROVIDER = "cosellProvider",
  AUTHORITY = "authority",
  BUDGET = "budget",
  TIME_LINE = "timeline",
  FILTER_STAGE = "filterStage",
  PURCHASE_INTENT = "purchaseIntent",
  NEED = "need",
  DECISION_PHASE = "decisionPhase",
  CONTRACT_VEHICLE = "contractVehicle",
  CONTRACT_VEHICLE_GCP = "contractVehicleGcp",
  REGION = "region",
  SUPPORT_LEVEL = "supportLevel",
  PRODUCT_FAMILY = "productFamily",
  PRODUCT_FAMILY_GCP = "productFamilyGcp",
  PARTNER_ENGAGEMENT = "partnerEngagementType",
  PARTNER_CONNECT = "partnerConnections",
  SELLER_CODE = "sellerCode",
  STAGE_AMP = "stageAmp",
  CLOSE_REASON = "ampClosingReason",
  SOURCE = "SourceGcp",
  STAGE_AWS = "stageAws",
  STAGE_GCP_CHANGE = "stageGcpChange",
}

export enum SortType {
  ASC = "ASC",
  DESC = "DESC",
}